ci: fix potential hang caused by prisma survey
add --no-hints to every prisma command to avoid
having this kind of survey.

```
How likely are you to recommend Prisma?

Enter a number from 0 to 10 (0 = not at all, 10 = extremely likely) and press
Enter — or leave blank to skip and not be asked again.

This prompt closes in 30s and can be suppressed with --no-hints. Learn more: https://pris.ly/why-nps
```